more jwt and security org

+++ src/main/java/com/TCorp/FitNetServer/api/repository/UserEntityRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserEntityRepository extends JpaRepository<UserEntity, Long> {
+public interface UserEntityRepository extends JpaRepository<UserEntity, Integer> {
     Optional<UserEntity> findUserEntityByEmailOrUsername(String email, String name);
 
     Optional<UserEntity> findUserEntityByUsername(String username);
